+++ src/cdm/DatabaseModel.ts
@@ -2,32 +2,40 @@ import { RowType } from "cdm/RowTypeModel"
 import { LocalSettings } from "Settings"
 
 /** database column */
-export type DatabaseColumn = {
-    input: string,
-    accessor: string,
-    label: string,
-    key?: string,
-    position?: number,
-    isMetadata?: boolean,
-    skipPersist?: boolean,
-    csvCandidate?: boolean,
-    isInline?: boolean,
-    [key: string]: RowType
+export interface DatabaseColumn {
+    input: string;
+    accessor: string;
+    label: string;
+    key: string;
+    position?: number;
+    isMetadata?: boolean;
+    skipPersist?: boolean;
+    csvCandidate?: boolean;
+    isInline?: boolean;
+    [key: string]: RowType;
 }
 
 /** database yaml */
-export type DatabaseYaml = {
+export interface DatabaseYaml {
     /** database name */
-    name: string,
+    name: string;
     /** database description */
-    description: string,
+    description: string;
     /** database columns */
-    columns: Record<string, DatabaseColumn>
+    columns: Record<string, DatabaseColumn>;
     /** database local configuration */
-    config?: LocalSettings
+    config: LocalSettings;
+    /** dataview filters */
+    filters?: FilterCondition[];
 }
 
 export type RowDatabaseFields = {
     frontmatter: Record<string, any>;
     inline: Record<string, any>;
+}
+
+export type FilterCondition = {
+    field: string;
+    operator: string;
+    value?: any;
 }