+++ src/components/HeaderMenu.tsx
@@ -4,7 +4,7 @@ import {
   TableDataType,
 } from "cdm/FolderModel";
 import { ActionTypes, DataTypes, StyleVariables } from "helpers/Constants";
-import { dbTrim } from "helpers/StylesHelper";
+import { dbTrim, c } from "helpers/StylesHelper";
 import ArrowUpIcon from "components/img/ArrowUp";
 import ArrowDownIcon from "components/img/ArrowDown";
 import ArrowLeftIcon from "components/img/ArrowLeft";
@@ -70,6 +70,8 @@ const HeaderMenu = (headerMenuProps: HeaderMenuProps) => {
   const [settingsPopperElement, setSettingsPopperElement] = useState(null);
   const [showSettings, setShowSettings] = useState(false);
 
+  // Manage errors
+  const [labelStateInvalid, setLabelStateInvalid] = useState(false);
   useEffect(() => {
     if (created) {
       setExpanded(true);
@@ -208,9 +210,7 @@ const HeaderMenu = (headerMenuProps: HeaderMenuProps) => {
     }
   );
 
-  function persistLabelChange() {
-    // trim label will get a valid yaml key
-    const newKey = dbTrim(labelState);
+  function persistLabelChange(newKey: string) {
     const futureOrder = headerMenuProps.headerProps.allColumns.map(
       (o: Column) => (o.id === column.id ? newKey : o.id)
     );
@@ -239,12 +239,26 @@ const HeaderMenu = (headerMenuProps: HeaderMenuProps) => {
   }
   function handleKeyDown(e: any) {
     if (e.key === "Enter") {
-      persistLabelChange();
+      // trim label will get a valid yaml key
+      const newKey = dbTrim(labelState);
+      // Check if key already exists. If so, mark it as invalid
+      if (
+        headerMenuProps.headerProps.allColumns.find(
+          (o: Column) => o.id === newKey
+        )
+      ) {
+        setLabelStateInvalid(true);
+        return;
+      }
+      persistLabelChange(newKey);
     }
   }
 
   function handleChange(e: any) {
     setLabelState(e.target.value);
+    if (labelStateInvalid) {
+      setLabelStateInvalid(false);
+    }
   }
 
   /**
@@ -290,10 +304,9 @@ const HeaderMenu = (headerMenuProps: HeaderMenuProps) => {
           {...attributes.popper}
         >
           <div
-            className="shadow-5 border-radius-md"
+            className={`menu ${c("popper")}`}
             style={{
               width: 240,
-              backgroundColor: StyleVariables.BACKGROUND_SECONDARY,
             }}
           >
             {/** Edit header label section */}
@@ -306,7 +319,11 @@ const HeaderMenu = (headerMenuProps: HeaderMenuProps) => {
             >
               <div className="is-fullwidth" style={{ marginBottom: 12 }}>
                 <input
-                  className="form-input"
+                  className={
+                    labelStateInvalid
+                      ? `${c("invalid-form")}`
+                      : `${c("form-input")}`
+                  }
                   ref={setInputRef}
                   type="text"
                   value={labelState}
@@ -328,9 +345,8 @@ const HeaderMenu = (headerMenuProps: HeaderMenuProps) => {
             </div>
             {/** Type of column section */}
             <div style={{ padding: "4px 0px" }}>
-              <button
-                className="sort-button"
-                type="button"
+              <div
+                className="menu-item sort-button"
                 onMouseEnter={() => setShowType(true)}
                 onMouseLeave={() => setShowType(false)}
                 ref={setTypeReferenceElement}
@@ -341,10 +357,10 @@ const HeaderMenu = (headerMenuProps: HeaderMenuProps) => {
                 <span style={{ textTransform: "capitalize" }}>
                   {column.dataType}
                 </span>
-              </button>
+              </div>
               {showType && (
                 <div
-                  className="shadow-5 border-radius-m"
+                  className={`menu ${c("popper")}`}
                   ref={setTypePopperElement}
                   onMouseEnter={() => setShowType(true)}
                   onMouseLeave={() => setShowType(false)}
@@ -352,19 +368,21 @@ const HeaderMenu = (headerMenuProps: HeaderMenuProps) => {
                   style={{
                     ...typePopper.styles.popper,
                     width: 200,
-                    backgroundColor: StyleVariables.BACKGROUND_SECONDARY,
                     zIndex: 4,
                     padding: "4px 0px",
                   }}
                 >
                   {types.map((type) => (
                     <div key={type.label}>
-                      <button className="sort-button" onClick={type.onClick}>
+                      <div
+                        className="menu-item sort-button"
+                        onClick={type.onClick}
+                      >
                         <span className="svg-icon svg-text icon-margin">
                           {type.icon}
                         </span>
                         {type.label}
-                      </button>
+                      </div>
                     </div>
                   ))}
                 </div>
@@ -373,35 +391,33 @@ const HeaderMenu = (headerMenuProps: HeaderMenuProps) => {
             {/** Action buttons section */}
             <div
               style={{
-                borderTop: `2px solid ${StyleVariables.BACKGROUND_DIVIDER}`,
+                borderTop: `1px solid ${StyleVariables.BACKGROUND_DIVIDER}`,
                 padding: "4px 0px",
               }}
             >
               {buttons.map((button) => (
-                <button
+                <div
                   key={button.label}
-                  type="button"
-                  className="sort-button"
+                  className="menu-item sort-button"
                   onMouseDown={button.onClick}
                 >
                   <span className="svg-icon svg-text icon-margin">
                     {button.icon}
                   </span>
                   {button.label}
-                </button>
+                </div>
               ))}
             </div>
             <div
               style={{
-                borderTop: `2px solid ${StyleVariables.BACKGROUND_DIVIDER}`,
+                borderTop: `1px solid ${StyleVariables.BACKGROUND_DIVIDER}`,
                 padding: "4px 0px",
               }}
             >
               {/** Column settings section */}
               <div style={{ padding: "4px 0px" }}>
-                <button
-                  className="sort-button"
-                  type="button"
+                <div
+                  className="menu-item sort-button"
                   onMouseEnter={() => setShowSettings(true)}
                   onMouseLeave={() => setShowSettings(false)}
                   ref={setSettingsReferenceElement}
@@ -410,10 +426,10 @@ const HeaderMenu = (headerMenuProps: HeaderMenuProps) => {
                     <AdjustmentsIcon />
                   </span>
                   <span>Settings</span>
-                </button>
+                </div>
                 {showSettings && (
                   <div
-                    className="shadow-5 border-radius-m"
+                    className={`menu ${c("popper")}`}
                     ref={setSettingsPopperElement}
                     onMouseEnter={() => setShowSettings(true)}
                     onMouseLeave={() => setShowSettings(false)}
@@ -421,7 +437,6 @@ const HeaderMenu = (headerMenuProps: HeaderMenuProps) => {
                     style={{
                       ...settingsPopper.styles.popper,
                       width: 200,
-                      backgroundColor: StyleVariables.BACKGROUND_SECONDARY,
                       zIndex: 4,
                       padding: "4px 0px",
                     }}