+++ src/parsers/RowDatabaseFieldsToFile.ts
@@ -1,6 +1,6 @@
 import { RowDatabaseFields } from "cdm/DatabaseModel";
 import { parseYaml } from "obsidian";
-export const parseFrontmatterFieldsToString = (databaseFields: RowDatabaseFields, original?: string, deletedColumn?: string): string => {
+export const parseFrontmatterFieldsToString = (databaseFields: RowDatabaseFields, original: string, deletedColumn?: string): string => {
     const frontmatterFields = databaseFields.frontmatter;
     const inlineFields = databaseFields.inline;
     const array: string[] = [];
@@ -8,19 +8,20 @@ export const parseFrontmatterFieldsToString = (databaseFields: RowDatabaseFields
     Object.keys(frontmatterFields).forEach(key => {
         array.push(`${key}: ${frontmatterFields[key]}`);
     });
-    if (original !== undefined) {
-        const match = original.match(/^---\s+([\w\W]+?)\s+---/);
-        if (match) {
-            const frontmatterRaw = match[1];
-            const yaml = parseYaml(frontmatterRaw);
-            Object.keys(yaml).forEach(key => {
+    const match = original.match(/^---\s+([\w\W]+?)\s+---/);
+    if (match) {
+        const frontmatterRaw = match[1];
+        const yaml = parseYaml(frontmatterRaw);
+        Object.keys(yaml)
+            .filter(fkey =>
+                // Filter out duplicates and deleted columns
+                !Object.prototype.hasOwnProperty.call(inlineFields, fkey)
+                && !Object.prototype.hasOwnProperty.call(frontmatterFields, fkey)
+                && fkey !== deletedColumn)
+            .forEach(key => {
                 // add frontmatter fields that are not specified as database fields
-                // check if frontmatter field is inside inline fields
-                if (!Object.prototype.hasOwnProperty.call(inlineFields, key) && !frontmatterFields[key] && key !== deletedColumn) {
-                    array.push(`${key}: ${yaml[key] ?? ''}`);
-                }
+                array.push(`${key}: ${yaml[key] ?? ''}`);
             });
-        }
     }
     array.push(`---`);
     return array.join('\n');